Action type Log for Microsoft_DefaultRuleSet_2.1
We are currently experiencing an issue when creating managed rule overrides whereby we cannot set the action type to Log for DRS version 2.1.

It is possible to set the action type to 'Log' or 'AnomalyScoring' via the portal and this is our current work around, but we need to be able to automate this fully. 

This proposed change is to update the validation to allow Log as a valid action type.